Exercises 4.15 - 4.12 done

+++ part_4/blogs/controllers/blogs.js
@@ -33,6 +33,7 @@ blogsRouter.post("/", async (request, response, next) => {
   const body = request.body;
   const token = getTokenFrom(request);
   const decodedToken = jwt.verify(token, process.env.SECRET);
+
   if (!decodedToken.id) {
     return response.status(401).json({ error: "token missing or invalid" });
   }
@@ -54,12 +55,24 @@ blogsRouter.post("/", async (request, response, next) => {
   response.json(savedBlog);
 });
 
-blogsRouter.delete("/:id", (request, response, next) => {
-  Blog.findByIdAndRemove(request.params.id)
-    .then(() => {
+blogsRouter.delete("/:id", async (request, response, next) => {
+  const token = request.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  const user = await User.findById(decodedToken.id);
+
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  if (blogToDelete.user._id.toString() === user._id.toString()) {
+    try {
+      await Blog.findByIdAndRemove(request.params.id);
       response.status(204).end();
-    })
-    .catch((error) => next(error));
+    } catch (exception) {
+      next(exception);
+    }
+  } else {
+    return response.status(401).json({ error: `Unauthorized` });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response, next) => {
